requirements.txt
einops
flit
h5py
joblib
joblib
jupyterlab
tensorboard
librosa
matplotlib
numpy==1.26.4
# onnxruntime-gpu
openpyxl # for saving results to excel
pandas
pesq
pre-commit
pystoi
pyyaml
scipy
soundfile
tabulate
toml
torchinfo
tqdm
accelerate
thop
huggingface_hub